Move getValueFromTheme
